requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extensions==2.2.9
idna==2.8
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.0
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
sqlparse==0.2.4
urllib3==1.24.2
psycopg2==2.8.5
six==1.15.0
gunicorn==20.0.4
django-heroku==0.3.1
whitenoise==5.1.0